backend/pyproject.toml
[build-system]
build-backend = "flit_core.buildapi"
requires = ["flit_core >=3.7.1,<4"]

[project]
name = "nettopp-flashcards-backend"
version = "0.1.0"
description = "The backend for the Nettop Flashcards project"
requires-python = ">=3.8"
classifiers = [ "Private :: Do Not Upload" ]
dependencies = [
    "alembic[tz]                ~= 1.9.1",
    "fastapi                    ~= 0.77.0",
    "psycopg2-binary            ~= 2.9.5",
    "python-dotenv              ~= 0.21.0",
    "python-multipart           ~= 0.0.5",
    "sqlalchemy-cockroachdb     ~= 1.4.4",
    "uvicorn                    ~= 0.17.0",

    # Requirements where the version depends on
    # other packages.
    "pydantic[email]",
]

[project.optional-dependencies]
lint = [
    "autopep8        ~= 2.0.1",
    "bandit[toml]    ~= 1.7.4",
    "flake8          ~= 6.0.0",
    "flake8-isort    ~= 6.0.0",
    "isort           ~= 5.10.0",
    "mypy            >= 0.950",
]
test = [
    "pytest          ~= 7.2.0",
    "pytest-cov      ~= 3.0.0",
    "pytest-asyncio  ~= 0.20.3",
]
dev = [
    "nox             ~= 2022.11.21",
    "pre-commit      ~= 2.20.0",
    "nettopp-flashcards-backend[test]",
    "nettopp-flashcards-backend[lint]",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-rA -q"
testpaths = [
    "tests"
]

[tool.flit.module]
name = "nfb"

[tool.bandit]
exclude_dirs = ["tests"]
